+++ site/view/content/git.php
@@ -1,21 +1,20 @@
 <!DOCTYPE html>
 <html lang="pt-BR">
 <head>
+    <?php
+    if ($_SERVER['SERVER_NAME'] != "localhost")
+        require '../includes/googleAnalytics.php';
+    ?>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
     <title>Git | Then Record</title>
 </head>
 <body>
-    
+        
     <?php 
         require '../../control/loadSheet.php';
-        
-        $sheetFile = "../../../dbSubjects/git/sheet/sheet-git.txt";
-        $commentsFile = "../../../dbSubjects/git/sheet/comments/pt-br-git.txt";
-        
-        loadSheetAndCommentWithLineNumber("&nbsp;&nbsp;&nbsp;", $sheetFile, "//", $commentsFile);
-            
+        displaySheet("git", " // ");
     ?>
 </body>
 </html>