+++ kanban/release-notes.md
@@ -1,5 +1,9 @@
 #Release Notes
 
+## 0.0.3
+[x] View database sheet content in front
+[x] Integration with Google Analytics 
+
 ## 0.0.2 
 [x] Control layer can receive any sheet view call